Shuffle around some code
